refactor: feature:gallery module package
