+++ feature/canvas/impl/src/main/java/com/teamyg/canvas/impl/route/CanvasImageAddRoute.kt
@@ -5,7 +5,7 @@ import androidx.compose.ui.Modifier
 import androidx.navigation3.runtime.result.ResultEffect
 import com.teamyg.camera.api.NavKeyCameraSystem
 import com.teamyg.canvas.impl.screen.CanvasImageAddScreen
-import com.teamyg.gallery.api.NavKeySystemGalleryPicker
+import com.teamyg.parfait.feature.gallery.api.NavKeySystemGalleryPicker
 import com.teamyg.navigation.Navigator
 import com.teamyg.parfait.feature.segmentation.api.NavKeySegmentation
