--- feature/gallery/api/src/main/java/com/teamyg/gallery/api/NavKeyCustomGalleryPicker.kt
@@ -1,7 +1,0 @@
-package com.teamyg.gallery.api
-
-import androidx.navigation3.runtime.NavKey
-import kotlinx.serialization.Serializable
-
-@Serializable
-data object NavKeyCustomGalleryPicker : NavKey
