+++ feature/gallery/api/src/main/java/com/teamyg/parfait/feature/gallery/api/NavKeySystemGalleryPicker.kt
@@ -1,0 +1,7 @@
+package com.teamyg.parfait.feature.gallery.api
+
+import androidx.navigation3.runtime.NavKey
+import kotlinx.serialization.Serializable
+
+@Serializable
+data object NavKeySystemGalleryPicker : NavKey
