--- feature/gallery/impl/src/main/java/com/teamyg/gallery/impl/model/GalleryAccessLevel.kt
@@ -1,34 +1,0 @@
-package com.teamyg.gallery.impl.model
-
-enum class GalleryAccessLevel {
-    INITIAL,
-    DENIED,
-    PERMANENTLY_DENIED,
-    PARTIAL,
-    FULL,
-    ;
-
-    val isPartial: Boolean
-        get() = when (this) {
-            PARTIAL -> true
-            else -> false
-        }
-
-    val hasPermission: Boolean
-        get() = when (this) {
-            PARTIAL,
-            FULL,
-            -> true
-
-            else -> false
-        }
-
-    val isDeniedPermission: Boolean
-        get() = when (this) {
-            DENIED,
-            PERMANENTLY_DENIED,
-            -> true
-
-            else -> false
-        }
-}
