--- feature/gallery/impl/src/main/java/com/teamyg/gallery/impl/model/GalleryAccessLevelPreviewParameterProvider.kt
@@ -1,13 +1,0 @@
-package com.teamyg.gallery.impl.model
-
-import androidx.compose.ui.tooling.preview.PreviewParameterProvider
-
-internal class GalleryAccessLevelPreviewParameterProvider : PreviewParameterProvider<GalleryAccessLevel> {
-    override val values: Sequence<GalleryAccessLevel>
-        get() = sequenceOf(
-            GalleryAccessLevel.DENIED,
-            GalleryAccessLevel.PERMANENTLY_DENIED,
-            GalleryAccessLevel.PARTIAL,
-            GalleryAccessLevel.FULL,
-        )
-}
