--- feature/gallery/impl/src/main/java/com/teamyg/gallery/impl/model/GalleryImageGroup.kt
@@ -1,9 +1,0 @@
-package com.teamyg.gallery.impl.model
-
-import androidx.compose.runtime.Immutable
-
-@Immutable
-data class GalleryImageGroup(
-    val date: String,
-    val images: List<String>,
-)
