+++ feature/gallery/impl/src/main/java/com/teamyg/parfait/feature/gallery/impl/model/GalleryAccessLevelPreviewParameterProvider.kt
@@ -1,0 +1,13 @@
+package com.teamyg.parfait.feature.gallery.impl.model
+
+import androidx.compose.ui.tooling.preview.PreviewParameterProvider
+
+internal class GalleryAccessLevelPreviewParameterProvider : PreviewParameterProvider<GalleryAccessLevel> {
+    override val values: Sequence<GalleryAccessLevel>
+        get() = sequenceOf(
+            GalleryAccessLevel.DENIED,
+            GalleryAccessLevel.PERMANENTLY_DENIED,
+            GalleryAccessLevel.PARTIAL,
+            GalleryAccessLevel.FULL,
+        )
+}
