+++ feature/gallery/impl/src/main/java/com/teamyg/parfait/feature/gallery/impl/model/GalleryImageGroup.kt
@@ -1,0 +1,9 @@
+package com.teamyg.parfait.feature.gallery.impl.model
+
+import androidx.compose.runtime.Immutable
+
+@Immutable
+data class GalleryImageGroup(
+    val date: String,
+    val images: List<String>,
+)
